tasks from phase 8
