+++ src/main/search/SearchView.tsx
@@ -1,16 +1,8 @@
-import { NetworkStatus } from '@apollo/client';
-
 import { Box, CircularProgress, Link } from '@mui/material';
-import { type PropsWithChildren, type ReactNode, useCallback, useEffect, useState } from 'react';
+import type { PropsWithChildren, ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
 import {
-  useSearchQuery,
-  useSearchScopeDetailsSpaceQuery,
-  useSpaceUrlResolverQuery,
-} from '@/core/apollo/generated/apollo-hooks';
-import {
   SearchCategory,
-  type SearchQuery,
   type SearchResult,
   type SearchResultCalloutFragment,
   type SearchResultMemoFragment,
@@ -21,8 +13,6 @@ import {
   type SearchResultUserFragment,
   type SearchResultWhiteboardFragment,
 } from '@/core/apollo/generated/graphql-schema';
-import useNavigate from '@/core/routing/useNavigate';
-import { useMemoizedQueryParams } from '@/core/routing/useQueryParams';
 import PageContentBlockSeamless from '@/core/ui/content/PageContentBlockSeamless';
 import PageContentColumn from '@/core/ui/content/PageContentColumn';
 import Gutters from '@/core/ui/grid/Gutters';
@@ -30,13 +20,10 @@ import { gutters } from '@/core/ui/grid/utils';
 import MultipleSelect from '@/core/ui/search/MultipleSelect';
 import SearchResultsScope from '@/core/ui/search/SearchResultsScope';
 import SearchResultsScopeCard from '@/core/ui/search/SearchResultsScopeCard';
-import { useCurrentUserContext } from '@/domain/community/userCurrent/useCurrentUserContext';
 import { SpaceL0Icon } from '@/domain/space/icons/SpaceL0Icon';
 import { buildLoginUrl } from '../routing/urlBuilders';
 import AlkemioLogo from '../ui/logo/logoSmall.svg?react';
-import { SEARCH_SPACE_URL_PARAM, SEARCH_TERMS_URL_PARAM } from './constants';
 import {
-  calloutFilterConfig,
   contributionFilterConfig,
   contributorFilterConfig,
   FILTER_OFF,
@@ -48,7 +35,7 @@ import SearchResultSection from './SearchResultSection';
 import SearchResultPostChooser from './searchResults/SearchResultPostChooser';
 import SearchResultsCalloutAndFramingCard from './searchResults/SearchResultsCalloutAndFramingCard';
 import SearchCategoriesMenu from './ui/SearchCategoriesMenu';
-import { useSearchTerms } from './useSearchTerms';
+import { useSearchViewState } from './useSearchViewState';
 
 export type TypedSearchResult<Type extends SearchResultType, ResultFragment extends {}> = SearchResult &
   ResultFragment & { type: Type };
@@ -69,27 +56,8 @@ interface SearchViewProps {
   spaceFilterTitle: ReactNode;
 }
 
-interface SearchViewSections {
-  spaceResults?: SearchResultMetaType[];
-  calloutResults?: SearchResultMetaType[];
-  framingResults?: SearchResultMetaType[]; // framingResults merged into calloutResults
-  contributionResults?: SearchResultMetaType[];
-  contributorResults?: SearchResultMetaType[];
-}
-
-type ResultsCursors = {
-  spaceCursor: string | undefined;
-  calloutCursor: string | undefined;
-  framingCursor: string | undefined;
-  contributionCursor: string | undefined;
-  contributorCursor: string | undefined;
-};
-
 const SEARCH_RESULTS_COUNT = 4;
 export const MAX_TERMS_SEARCH = 5;
-const tagsetNames = ['skills', 'keywords'];
-
-const concatSearchResults = <T,>(a: T[] = [], b: T[] = []): T[] => [...a, ...b];
 
 const interlaceAndFilterArrays = <T extends { type: SearchResultType }>(
   a: T[] = [],
@@ -115,394 +83,40 @@ const interlaceAndFilterArrays = <T extends { type: SearchResultType }>(
 const Logo = () => <AlkemioLogo />;
 
 const SearchView = ({ searchRoute, spaceFilterConfig, spaceFilterTitle }: SearchViewProps) => {
-  const [canSpaceLoadMore, setCanSpaceLoadMore] = useState(true);
-  const [canCalloutLoadMore, setCanCalloutLoadMore] = useState(true);
-  const [canFramingLoadMore, setCanFramingLoadMore] = useState(true);
-  const [canContributionLoadMore, setCanContributionLoadMore] = useState(true);
-  const [canContributorLoadMore, setContributorCanLoadMore] = useState(true);
-
-  const [resultsCursors, setResultsCursors] = useState<ResultsCursors>({
-    spaceCursor: undefined,
-    calloutCursor: undefined,
-    framingCursor: undefined,
-    contributionCursor: undefined,
-    contributorCursor: undefined,
-  });
-  const [spaceFilter, setSpaceFilter] = useState<FilterDefinition>(spaceFilterConfig.all);
-  const [calloutFilter, setCalloutFilter] = useState<FilterDefinition>(calloutFilterConfig.all);
-  const [framingFilter, setFramingFilter] = useState<FilterDefinition>(framingFilterConfig.all);
-  const [contributionFilter, setContributionFilter] = useState<FilterDefinition>(contributionFilterConfig.all);
-  const [contributorFilter, setContributorFilter] = useState<FilterDefinition>(contributorFilterConfig.all);
-
-  const navigate = useNavigate();
-
   const { t } = useTranslation();
-
-  const termsFromUrl = useSearchTerms();
-
-  const queryParams = useMemoizedQueryParams();
-
-  const { isAuthenticated } = useCurrentUserContext();
-
-  const spaceNameId = queryParams[SEARCH_SPACE_URL_PARAM]?.[0] ?? undefined;
-
-  const handleTermsChange = useCallback(
-    (newValue: string[]) => {
-      setCanSpaceLoadMore(true);
-      setCanCalloutLoadMore(true);
-      setCanFramingLoadMore(true);
-      setCanContributionLoadMore(true);
-      setContributorCanLoadMore(true);
-
-      const params = new URLSearchParams(
-        Object.entries(queryParams).flatMap(([key, values]) => values.map(value => [key, value]))
-      );
-
-      params.delete(SEARCH_TERMS_URL_PARAM);
-
-      for (const term of newValue) {
-        params.append(SEARCH_TERMS_URL_PARAM, term);
-      }
-
-      if (newValue.length === 0) {
-        params.append(SEARCH_TERMS_URL_PARAM, '');
-      }
-
-      navigate(`${searchRoute}?${params}`);
-    },
-    [searchRoute, queryParams, navigate]
-  );
-
-  const handleSearchInPlatform = useCallback(() => {
-    const params = new URLSearchParams(
-      Object.entries(queryParams).flatMap(([key, values]) => values.map(value => [key, value]))
-    );
-
-    params.delete(SEARCH_SPACE_URL_PARAM);
-
-    navigate(`${searchRoute}?${params}`);
-  }, [searchRoute, queryParams, navigate]);
-
-  const { data: spaceIdData, loading: resolvingSpace } = useSpaceUrlResolverQuery({
-    variables: { spaceNameId: spaceNameId! },
-    skip: !spaceNameId,
-  });
-
-  const spaceId = spaceIdData?.lookupByName.space?.id;
-
-  const hasNoTermsLength = termsFromUrl.length === 0;
 
   const {
     data,
-    networkStatus,
-    loading: isSearching,
-    fetchMore,
-  } = useSearchQuery({
-    variables: {
-      searchData: {
-        tagsetNames,
-        terms: termsFromUrl,
-        searchInSpaceFilter: spaceId,
-        filters: [
-          {
-            category: SearchCategory.Spaces,
-            size: SEARCH_RESULTS_COUNT,
-            types: spaceFilterConfig.all.value,
-            cursor: undefined,
-          },
-          {
-            category: SearchCategory.CollaborationTools,
-            size: SEARCH_RESULTS_COUNT,
-            types: calloutFilterConfig.all.value,
-            cursor: undefined,
-          },
-          {
-            category: SearchCategory.Framings,
-            size: SEARCH_RESULTS_COUNT,
-            types: framingFilterConfig.all.value,
-            cursor: undefined,
-          },
-          {
-            category: SearchCategory.Contributions,
-            size: SEARCH_RESULTS_COUNT,
-            types: contributionFilterConfig.all.value,
-            cursor: undefined,
-          },
-          {
-            category: SearchCategory.Contributors,
-            size: SEARCH_RESULTS_COUNT,
-            types: contributorFilterConfig.all.value,
-            cursor: undefined,
-          },
-        ],
-      },
-    },
-    fetchPolicy: 'no-cache',
-    skip: hasNoTermsLength || resolvingSpace,
-  });
-
-  const { spaceResults, calloutResults, framingResults, contributionResults, contributorResults }: SearchViewSections =
-    toResultType(hasNoTermsLength ? undefined : data);
-
-  const { data: spaceDetails, loading } = useSearchScopeDetailsSpaceQuery({
-    variables: { spaceId: spaceId! },
-    skip: !spaceId,
-  });
-
-  const fetchNewResults = useCallback(
-    (resultsType: SearchCategory) => {
-      const getFetchFilters = () => {
-        switch (resultsType) {
-          case SearchCategory.Spaces: {
-            return [
-              {
-                category: SearchCategory.Spaces,
-                size: SEARCH_RESULTS_COUNT,
-                types: spaceFilter.value,
-                cursor: resultsCursors.spaceCursor,
-              },
-            ];
-          }
-
-          case SearchCategory.CollaborationTools: {
-            return [
-              {
-                category: SearchCategory.CollaborationTools,
-                size: SEARCH_RESULTS_COUNT,
-                // types: calloutFilter.value,  // Callouts do not have type filter
-                cursor: resultsCursors.calloutCursor,
-              },
-            ];
-          }
-
-          case SearchCategory.Framings: {
-            return [
-              {
-                category: SearchCategory.Framings,
-                size: SEARCH_RESULTS_COUNT,
-                types: framingFilter.value,
-                cursor: resultsCursors.framingCursor,
-              },
-            ];
-          }
-
-          case SearchCategory.Contributions: {
-            return [
-              {
-                category: SearchCategory.Contributions,
-                size: SEARCH_RESULTS_COUNT,
-                types: contributionFilter.value,
-                cursor: resultsCursors.contributionCursor,
-              },
-            ];
-          }
-
-          case SearchCategory.Contributors: {
-            return [
-              {
-                category: SearchCategory.Contributors,
-                size: SEARCH_RESULTS_COUNT,
-                types: contributorFilter.value,
-                cursor: resultsCursors.contributorCursor,
-              },
-            ];
-          }
-
-          default: {
-            return [
-              {
-                category: SearchCategory.Spaces,
-                size: SEARCH_RESULTS_COUNT,
-                types: spaceFilter.value,
-                cursor: resultsCursors.spaceCursor,
-              },
-              {
-                category: SearchCategory.CollaborationTools,
-                size: SEARCH_RESULTS_COUNT,
-                // types: calloutFilter.value,  // Callouts do not have type filter
-                cursor: resultsCursors.calloutCursor,
-              },
-              {
-                category: SearchCategory.Framings,
-                size: SEARCH_RESULTS_COUNT,
-                types: framingFilter.value,
-                cursor: resultsCursors.framingCursor,
-              },
-              {
-                category: SearchCategory.Contributions,
-                size: SEARCH_RESULTS_COUNT,
-                types: contributionFilter.value,
-                cursor: resultsCursors.contributionCursor,
-              },
-              {
-                category: SearchCategory.Contributors,
-                size: SEARCH_RESULTS_COUNT,
-                types: contributorFilter.value,
-                cursor: resultsCursors.contributorCursor,
-              },
-            ];
-          }
-        }
-      };
-
-      fetchMore({
-        variables: {
-          searchData: {
-            tagsetNames,
-            terms: termsFromUrl,
-            filters: getFetchFilters(),
-            searchInSpaceFilter: spaceId,
-          },
-        },
-
-        updateQuery: (prev: SearchQuery, { fetchMoreResult }: { fetchMoreResult: SearchQuery }) => {
-          switch (resultsType) {
-            case SearchCategory.Spaces: {
-              setCanSpaceLoadMore(fetchMoreResult?.search?.spaceResults?.results?.length > 0);
-              break;
-            }
-
-            case SearchCategory.CollaborationTools: {
-              setCanCalloutLoadMore(fetchMoreResult?.search?.calloutResults?.results?.length > 0);
-              break;
-            }
-            case SearchCategory.Framings: {
-              setCanFramingLoadMore(fetchMoreResult?.search?.framingResults?.results?.length > 0);
-              break;
-            }
-
-            case SearchCategory.Contributions: {
-              setCanContributionLoadMore(fetchMoreResult?.search?.contributionResults?.results?.length > 0);
-              break;
-            }
-
-            case SearchCategory.Contributors: {
-              setContributorCanLoadMore(fetchMoreResult?.search?.actorResults?.results?.length > 0);
-              break;
-            }
-
-            default: {
-              break;
-            }
-          }
-
-          switch (resultsType) {
-            case SearchCategory.Spaces: {
-              return {
-                search: {
-                  ...prev.search,
-                  spaceResults: {
-                    ...fetchMoreResult.search.spaceResults,
-                    results: concatSearchResults(
-                      prev.search.spaceResults?.results,
-                      fetchMoreResult.search.spaceResults?.results
-                    ),
-                  },
-                },
-              };
-            }
-
-            case SearchCategory.CollaborationTools: {
-              return {
-                search: {
-                  ...prev.search,
-                  calloutResults: {
-                    ...fetchMoreResult.search.calloutResults,
-                    results: concatSearchResults(
-                      prev.search.calloutResults?.results,
-                      fetchMoreResult.search.calloutResults?.results
-                    ),
-                  },
-                },
-              };
-            }
-
-            case SearchCategory.Framings: {
-              return {
-                search: {
-                  ...prev.search,
-                  framingResults: {
-                    ...fetchMoreResult.search.framingResults,
-                    results: concatSearchResults(
-                      prev.search.framingResults?.results,
-                      fetchMoreResult.search.framingResults?.results
-                    ),
-                  },
-                },
-              };
-            }
-
-            case SearchCategory.Contributions: {
-              return {
-                search: {
-                  ...prev.search,
-                  contributionResults: {
-                    ...fetchMoreResult.search.contributionResults,
-                    results: concatSearchResults(
-                      prev.search.contributionResults?.results,
-                      fetchMoreResult.search.contributionResults?.results
-                    ),
-                  },
-                },
-              };
-            }
-
-            case SearchCategory.Contributors: {
-              return {
-                search: {
-                  ...prev.search,
-                  actorResults: {
-                    ...fetchMoreResult.search.actorResults,
-                    results: concatSearchResults(
-                      prev.search.actorResults?.results,
-                      fetchMoreResult.search.actorResults?.results
-                    ),
-                  },
-                },
-              };
-            }
-
-            default:
-              return prev;
-          }
-        },
-      });
-    },
-    [
-      spaceId,
-      termsFromUrl,
-      resultsCursors,
-      spaceFilter.value,
-      calloutFilter.value,
-      framingFilter.value,
-      contributionFilter.value,
-      contributorFilter.value,
-      fetchMore,
-    ]
-  );
-
-  useEffect(() => {
-    if (hasNoTermsLength) {
-      setSpaceFilter(spaceFilterConfig.all);
-      setCalloutFilter(calloutFilterConfig.all);
-      setFramingFilter(framingFilterConfig.all);
-      setContributionFilter(contributionFilterConfig.all);
-      setContributorFilter(contributorFilterConfig.all);
-    }
-  }, [hasNoTermsLength]);
-
-  useEffect(() => {
-    if (data?.search && !isSearching)
-      setResultsCursors({
-        spaceCursor: data.search.spaceResults?.cursor ?? undefined, // This check is required since the BE return `null` when cursor is missing
-        calloutCursor: data.search.calloutResults?.cursor ?? undefined, // This check is required since the BE return `null` when cursor is missing
-        framingCursor: data.search.framingResults?.cursor ?? undefined, // This check is required since the BE return `null` when cursor is missing
-        contributionCursor: data.search.contributionResults?.cursor ?? undefined, // This check is required since the BE return `null` when cursor is missing
-        contributorCursor: data.search.actorResults?.cursor ?? undefined, // This check is required since the BE return `null` when cursor is missing
-      });
-  }, [data, isSearching]);
-
-  const isSearchingForMore = networkStatus === NetworkStatus.fetchMore;
+    termsFromUrl,
+    isAuthenticated,
+    isSearching,
+    isSearchingForMore,
+    hasNoTermsLength,
+    spaceId,
+    spaceDetails,
+    spaceDetailsLoading: loading,
+    spaceResults,
+    calloutResults,
+    framingResults,
+    contributionResults,
+    contributorResults,
+    spaceFilter,
+    setSpaceFilter,
+    framingFilter,
+    setFramingFilter,
+    contributionFilter,
+    setContributionFilter,
+    contributorFilter,
+    setContributorFilter,
+    canSpaceLoadMore,
+    canCalloutLoadMore,
+    canFramingLoadMore,
+    canContributionLoadMore,
+    canContributorLoadMore,
+    handleTermsChange,
+    handleSearchInPlatform,
+    fetchNewResults,
+  } = useSearchViewState(searchRoute, spaceFilterConfig);
 
   const filteredSpaceResults =
     spaceFilter.typename === 'all'
@@ -681,36 +295,6 @@ const SearchView = ({ searchRoute, spaceFilterConfig, spaceFilterTitle }: Search
 };
 
 export default SearchView;
-
-function toResultType(query?: SearchQuery): SearchViewSections {
-  const spaceResults = (query?.search.spaceResults?.results || []).map<SearchResultMetaType>(
-    ({ score, terms, ...rest }) => ({ ...rest, score: score || 0, terms: terms || [] }) as SearchResultMetaType
-  );
-
-  const calloutResults = (query?.search.calloutResults?.results || []).map<SearchResultMetaType>(
-    ({ score, terms, ...rest }) => ({ ...rest, score: score || 0, terms: terms || [] }) as SearchResultMetaType
-  );
-
-  const framingResults = (query?.search.framingResults?.results || []).map<SearchResultMetaType>(
-    ({ score, terms, ...rest }) => ({ ...rest, score: score || 0, terms: terms || [] }) as SearchResultMetaType
-  );
-
-  const contributionResults = (query?.search.contributionResults?.results || []).map<SearchResultMetaType>(
-    ({ score, terms, ...rest }) => ({ ...rest, score: score || 0, terms: terms || [] }) as SearchResultMetaType
-  );
-
-  const contributorResults = (query?.search.actorResults?.results || []).map<SearchResultMetaType>(
-    ({ score, terms, ...rest }) => ({ ...rest, score: score || 0, terms: terms || [] }) as SearchResultMetaType
-  );
-
-  return {
-    spaceResults,
-    calloutResults,
-    framingResults,
-    contributionResults,
-    contributorResults,
-  };
-}
 
 function SectionWrapper({ children }: PropsWithChildren) {
   return <Box sx={{ display: 'flex', flexDirection: 'row', gap: gutters(1) }}>{children}</Box>;
